+++ winsnap/collectors/__init__.py
@@ -2,6 +2,7 @@ from winsnap.collectors.processes import collect_processes
 from winsnap.collectors.registry_autoruns import collect_registry_autoruns
 from winsnap.collectors.scheduled_tasks import collect_scheduled_tasks
 from winsnap.collectors.services import collect_services
+from winsnap.collectors.startup_folders import collect_startup_folders
 
 
 __all__ = [
@@ -9,4 +10,5 @@ __all__ = [
     "collect_registry_autoruns",
     "collect_scheduled_tasks",
     "collect_services",
+    "collect_startup_folders",
 ]